Utility: Move string functions to lib `kdbutility`
