+++ src/include/CMakeLists.txt
@@ -74,6 +74,7 @@ install (FILES
 	kdbplugin.h
 	kdbprivate.h
 	kdbproposal.h
+	kdbutility.h
 	"${CMAKE_CURRENT_BINARY_DIR}/kdbversion.h"
 	DESTINATION include/${TARGET_INCLUDE_FOLDER})
