+++ src/include/kdbprivate.h
@@ -553,9 +553,4 @@ void elektraGlobalError (KDB * handle, KeySet * ks, Key * parentKey, int positio
 }
 #endif
 
-/* Convenience Methods for String Handling */
-char * lskip (char const * const keyname);
-char * rstrip (char * const start, char ** end);
-char * strip (char * text);
-
 #endif /* KDBPRIVATE_H */
