+++ src/libs/CMakeLists.txt
@@ -12,6 +12,8 @@ add_subdirectory (meta)
 
 add_subdirectory (plugin)
 
+add_subdirectory (utility)
+
 add_subdirectory (elektra)
 
 generate_manpage (elektra-libs FILENAME ${CMAKE_CURRENT_SOURCE_DIR}/README.md SECTION 7)
