+++ src/libs/utility/CMakeLists.txt
@@ -1,0 +1,2 @@
+file (GLOB SOURCES *.c)
+add_lib (utility SOURCES ${SOURCES})
